+++ packages/aptos-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @irys/upload-aptos
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @irys/upload@0.0.8
+
 ## 0.0.7
 
 ### Patch Changes