+++ packages/solana-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @irys/upload-solana
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @irys/upload@0.0.8
+
 ## 0.1.0
 
 ### Minor Changes