+++ packages/solana-web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @irys/web-upload-solana
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @irys/web-upload@0.0.8
+
 ## 0.1.0
 
 ### Minor Changes